refactor(analysis): extract content cyclomatic scoring
